#1 - 임시 Travel struct 생성(추후 삭제)
