+++ TravelTale/Common/Travel(temporary).swift
@@ -1,0 +1,16 @@
+//
+//  Travel.swift
+//  TravelTale
+//
+//  Created by 유림 on 6/5/24.
+//
+
+import Foundation
+
+struct Travel {
+    let image: Data?
+    let title: String
+    let startDate: Date
+    let endDate: Date
+    let province: String?
+}
