+++ TravelTale/Views/Travels/TravelViewModel.swift
@@ -7,17 +7,6 @@
 
 import Foundation
 
-// ------------------------------------
-// 임시 데이터 구조
-struct Travel {
-    let image: Data?
-    let title: String
-    let startDate: Date
-    let endDate: Date
-    let province: String?
-}
-// ------------------------------------
-
 class TravelViewModel {
     
     var travelArray: Observable<[Travel]> = .init([])
